apps/EZWorking/src/services/camel-agents/requirements.txt
camel-ai>=0.1.5
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
asyncio-mqtt>=0.13.0
redis>=5.0.0